dom/plugins/test/mochitest/test_npn_asynccall.html

<html>
<head>
  <title>NPN_AsyncCallback Tests</title>
  <script type="text/javascript" 
          src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" 
        href="/tests/SimpleTest/test.css" />
</head>
<body onload="runTests()">
  <p id="display"></p>

  <embed id="plugin1" type="application/x-test" width="400" height="100">
  </embed>

  <script class="testbody" type="application/javascript">
  
  SimpleTest.waitForExplicitFinish();
  
  function asyncTestsComplete(result) {
    ok(result, "asyncCallback tests completed");
    SimpleTest.finish();
  }
  
  function runTests() {
    var plugin = document.getElementById("plugin1");
    plugin.asyncCallbackTest("asyncTestsComplete");
  }
  </script>
  
 </body>
 </html>
